+++ immutable-arrays/core/src/main/kotlin/com/danrusu/pods4k/immutableArrays/ImmutableFloatArray.kt
@@ -312,6 +312,21 @@ public value class ImmutableFloatArray @PublishedApi internal constructor(
     }
 
     /**
+     * Returns an immutable array containing the last [n] elements.
+     *
+     * @throws IllegalArgumentException if [n] is negative.
+     */
+    public fun takeLast(n: Int): ImmutableFloatArray {
+        require(n >= 0) { "Requested element count $n is less than zero." }
+        if (n == 0) return EMPTY
+        if (n >= size) return this
+
+        val backingArray = FloatArray(n)
+        System.arraycopy(values, size - n, backingArray, 0, n)
+        return ImmutableFloatArray(backingArray)
+    }
+
+    /**
      * Returns an immutable array containing only the elements matching the given [predicate].
      */
     public inline fun filter(predicate: (element: Float) -> Boolean): ImmutableFloatArray {